part 1

+++ src/Client.java
@@ -18,19 +18,19 @@ public class Client {
         System.out.println(this.time);
 
         try (Socket socket = new Socket(hostname, port); OutputStream out = socket.getOutputStream()) {
-            byte[] data = new byte[1000]; // Chunk of 1000 bytes
+            byte[] data = new byte[1000];
             long startTime = System.currentTimeMillis();
             long endTime = startTime + (time * 1000);
-            int totalBytesSent = 0;
+            long totalBytesSent = 0;
             
             while (System.currentTimeMillis() < endTime) {
                 out.write(data);
                 totalBytesSent += data.length;
             }
 
-            double sentKB = totalBytesSent / 1000.0;
+            long sentKB = totalBytesSent / 1000;
             double sentMbps = (totalBytesSent * 8.0) / (time * 1000000.0);
-            System.out.printf("sent=%.3f KB rate=%.3f Mbps%n", sentKB, sentMbps);
+            System.out.printf("sent=%d KB rate=%.3f Mbps%n", sentKB, sentMbps);
         } catch (Exception e) {
             System.err.println("Error: " + e.getMessage());
         }
